+++ src/tinydiffusion/utils/__init__.py
@@ -2,5 +2,12 @@
 
 from tinydiffusion.utils.modules import eval_mode
 from tinydiffusion.utils.seed import seed_everything
+from tinydiffusion.utils.tracking import RunLogger, null_logger, timestep_quartile_losses
 
-__all__ = ["eval_mode", "seed_everything"]
+__all__ = [
+    "RunLogger",
+    "eval_mode",
+    "null_logger",
+    "seed_everything",
+    "timestep_quartile_losses",
+]